fix: Make golangci-lint rules more strict and fix new linter warnings

+++ pkg/capture.go
@@ -128,7 +128,7 @@ func NewCapture(opts ...CaptureOption) *Capture {
 
 // Count returns the total number of captured packets
 func (c *Capture) Count() uint64 {
-	return uint64(c.count.Load())
+	return c.count.Load()
 }
 
 func (c *Capture) Flush() error {
